+++ src/shaders/glsl/symbol_sdf.vertex.glsl
@@ -3,7 +3,7 @@ layout(location = 1) in uvec4 a_data;
 layout(location = 2) in vec4 a_pixeloffset;
 layout(location = 3) in vec3 a_projected_pos;
 layout(location = 4) in uint a_fade_opacity;
-layout(location = 5) in float a_elevation;
+layout(location = 5) in float a_height_offset;
 
 // contents of a_size vary based on the type of property value
 // used for {text,icon}-size.
@@ -31,6 +31,7 @@ uniform vec2 u_texsize;
 uniform vec2 u_translation;
 uniform float u_pitched_scale;
 uniform bool u_is_offset;
+uniform bool u_height_anchor_ground;
 
 out vec2 v_data0;
 out vec3 v_data1;
@@ -58,7 +59,7 @@ void main() {
     vec2 a_pxoffset = a_pixeloffset.xy / 16.0;
     vec2 a_minFontScale = a_pixeloffset.zw / 256.0;
 
-    float ele = get_elevation(a_pos) + a_elevation;
+    float ele = a_height_offset + (u_height_anchor_ground ? get_elevation(a_pos) : 0.0);
     highp float segment_angle = -a_projected_pos[2];
     float size;