+++ src/shaders/glsl/symbol_text_and_icon.vertex.glsl
@@ -2,7 +2,7 @@ layout(location = 0) in vec4 a_pos_offset;
 layout(location = 1) in uvec4 a_data;
 layout(location = 2) in vec3 a_projected_pos;
 layout(location = 3) in uint a_fade_opacity;
-layout(location = 4) in float a_elevation;
+layout(location = 4) in float a_height_offset;
 
 // contents of a_size vary based on the type of property value
 // used for {text,icon}-size.
@@ -31,6 +31,7 @@ uniform bool u_is_variable_anchor;
 uniform vec2 u_translation;
 uniform float u_pitched_scale;
 uniform bool u_is_offset;
+uniform bool u_height_anchor_ground;
 
 out vec4 v_data0;
 out vec3 v_data1;
@@ -58,7 +59,7 @@ void main() {
     float a_size_min = float(a_data.z >> 1u);
     float is_sdf = float(a_data.z & 1u);
 
-    float ele = get_elevation(a_pos) + a_elevation;
+    float ele = a_height_offset + (u_height_anchor_ground ? get_elevation(a_pos) : 0.0);
     highp float segment_angle = -a_projected_pos[2];
     float size;